Use select query to test reading behavior
